Latest version
The project aims to develop a simple machine learning model that is able to estimate house prices based on a set of factors and variables that affect the value of the property. Data from the Kaggle platform was used to train the model, which is a simple model that can be used by real estate companies, individuals, or financial institutions to value properties.